+++ SandTetris/Interfaces/ISalaryService.cs
@@ -3,8 +3,8 @@ namespace SandTetris.Interfaces;
 public interface ISalaryService
 {
     Task CalculateSalariesAsync(string departmentId, int month, int year);
-    Task<decimal> GetEmployeeSalaryAsync(string employeeId, int month, int year);
-    Task<decimal> GetDepartmentSalaryAsync(string departmentId, int month, int year);
-    Task<decimal> GetTotalSalaryAsync(int month, int year);
-    Task<decimal> GetTotalAll();
+    Task<int> GetEmployeeSalaryAsync(string employeeId, int month, int year);
+    Task<int> GetDepartmentSalaryAsync(string departmentId, int month, int year);
+    Task<int> GetTotalSalaryAsync(int month, int year);
+    Task<int> GetTotalAll();
 }
